refactor: 패키지명 오타(caculator → calculator) 정정 및 전체 경로 일괄 수정
