+++ src/main/java/calculator/domain/CalculationSequence.java
@@ -1,4 +1,4 @@
-package caculator.domain;
+package calculator.domain;
 
 import java.util.List;
