+++ src/main/java/calculator/domain/Expression.java
@@ -1,4 +1,4 @@
-package caculator.domain;
+package calculator.domain;
 
 import java.util.List;
